partially move mid and mtrust to own modules
