+++ packages/kuma-gui/src/app/mesh-identity/index.ts
@@ -1,0 +1,21 @@
+import { token } from '@kumahq/container'
+
+import { sources } from './sources'
+import type { ServiceDefinition } from '@kumahq/container'
+
+type Token = ReturnType<typeof token>
+type ResourcesSources = ReturnType<typeof sources>
+
+export const services = (app: Record<string, Token>): ServiceDefinition[] => {
+  return [
+    [token<ResourcesSources>('mesh-insight.sources'), {
+      service: sources,
+      arguments: [
+        app.api,
+      ],
+      labels: [
+        app.sources,
+      ],
+    }],
+  ]
+}
