+++ packages/kuma-gui/src/app/meshes/index.ts
@@ -12,6 +12,8 @@ import { services as dataplanes } from '@/app/data-planes'
 import { services as externalServicesModule } from '@/app/external-services'
 import { services as gatewaysModule } from '@/app/gateways'
 import { services as legacyDataplanes } from '@/app/legacy-data-planes'
+import { services as meshIdentity } from '@/app/mesh-identity'
+import { services as meshTrust } from '@/app/mesh-trust'
 import { services as policies } from '@/app/policies'
 import { services as resources } from '@/app/resources'
 import { services as rules } from '@/app/rules'
@@ -96,6 +98,8 @@ export const services = (app: Record<string, Token>): ServiceDefinition[] => {
     ...rules(mesh),
     ...workloads(mesh),
     ...resources(mesh),
+    ...meshIdentity(mesh),
+    ...meshTrust(mesh),
   ]
 }
 export const TOKENS = $
