+++ packages/kuma-gui/src/app/resources/sources.ts
@@ -1,10 +1,6 @@
 import createClient from 'openapi-fetch'
 
 import { Resource } from './data'
-import { MeshIdentity } from './data/MeshIdentity'
-import type { KumaMeshIdentity } from './data/MeshIdentity'
-import { MeshTrust } from './data/MeshTrust'
-import type { KumaMeshTrust } from './data/MeshTrust'
 import { Resources } from './data/Resources'
 import { ResourcesTypes } from './data/ResourceTypes'
 import { defineSources } from '@/app/application'
@@ -94,98 +90,6 @@ export const sources = (api: KumaApi) => {
       })
 
       return response.data!
-    },
-
-    '/meshes/:mesh/meshidentities': async (params) => {
-      const { mesh } = params
-  
-      const res = await http.GET('/meshes/{mesh}/meshidentities', {
-        params: {
-          path: {
-            mesh,
-          },
-        },
-      })
-  
-      return MeshIdentity.fromCollection(res.data!)
-    },
-
-    '/meshidentities/:mid': async (params) => {
-      const { mid } = params
-  
-      const res = await http.GET('/_kri/{kri}', {
-        params: {
-          path: {
-            kri: mid,
-          },
-        },
-      })
-  
-      return MeshIdentity.fromObject(res.data as KumaMeshIdentity)
-    },
-
-    '/meshidentities/:mid/as/kubernetes': async (params) => {
-      const { mid } = params
-  
-      const res = await http.GET('/_kri/{kri}', {
-        params: {
-          path: {
-            kri: mid,
-          },
-          // @ts-ignore
-          query: {
-            format: 'kubernetes',
-          },
-        },
-      })
-  
-      return res.data
-    },
-
-    '/meshes/:mesh/meshtrusts': async (params) => {
-      const { mesh } = params
-  
-      const res = await http.GET('/meshes/{mesh}/meshtrusts', {
-        params: {
-          path: {
-            mesh,
-          },
-        },
-      })
-  
-      return MeshTrust.fromCollection(res.data!)
-    },
-
-    '/meshtrusts/:mtrust': async (params) => {
-      const { mtrust } = params
-  
-      const res = await http.GET('/_kri/{kri}', {
-        params: {
-          path: {
-            kri: mtrust,
-          },
-        },
-      })
-  
-      return MeshTrust.fromObject(res.data as KumaMeshTrust)
-    },
-
-    '/meshtrusts/:mtrust/as/kubernetes': async (params) => {
-      const { mtrust } = params
-  
-      const res = await http.GET('/_kri/{kri}', {
-        params: {
-          path: {
-            kri: mtrust,
-          },
-          // @ts-ignore
-          query: {
-            format: 'kubernetes',
-          },
-        },
-      })
-  
-      return res.data
     },
   })
 }
